+++ tower-hedge/tests/hedge.rs
@@ -1,154 +1,135 @@
-use futures_util::pin_mut;
-use std::future::Future;
 use std::time::Duration;
-use tokio_test::{assert_pending, assert_ready, assert_ready_ok, clock, task};
+use tokio::time;
+use tokio_test::{assert_pending, assert_ready, assert_ready_ok, task};
 use tower_hedge::{Hedge, Policy};
-use tower_service::Service;
-use tower_test::assert_request_eq;
+use tower_test::{assert_request_eq, mock};
 
-#[test]
-fn hedge_orig_completes_first() {
-    task::mock(|cx| {
-        clock::mock(|time| {
-            let (mut service, handle) = new_service(TestPolicy);
-            pin_mut!(handle);
+#[tokio::test]
+async fn hedge_orig_completes_first() {
+    time::pause();
 
-            assert_ready_ok!(service.poll_ready(cx));
-            let fut = service.call("orig");
-            pin_mut!(fut);
+    let (mut service, mut handle) = new_service(TestPolicy);
 
-            // Check that orig request has been issued.
-            let req = assert_request_eq!(handle, "orig");
-            // Check fut is not ready.
-            assert_pending!(fut.as_mut().poll(cx));
+    assert_ready_ok!(service.poll_ready());
+    let mut fut = task::spawn(service.call("orig"));
 
-            // Check hedge has not been issued.
-            assert_pending!(handle.as_mut().poll_request(cx));
-            time.advance(Duration::from_millis(10));
-            // Check fut is not ready.
-            assert_pending!(fut.as_mut().poll(cx));
-            // Check that the hedge has been issued.
-            let _hedge_req = assert_request_eq!(handle, "orig");
+    // Check that orig request has been issued.
+    let req = assert_request_eq!(handle, "orig");
+    // Check fut is not ready.
+    assert_pending!(fut.poll());
 
-            req.send_response("orig-done");
-            // Check that fut gets orig response.
-            assert_eq!(assert_ready_ok!(fut.as_mut().poll(cx)), "orig-done");
-        });
-    });
+    // Check hedge has not been issued.
+    assert_pending!(handle.poll_request());
+    time::advance(Duration::from_millis(11)).await;
+    // Check fut is not ready.
+    assert_pending!(fut.poll());
+    // Check that the hedge has been issued.
+    let _hedge_req = assert_request_eq!(handle, "orig");
+
+    req.send_response("orig-done");
+    // Check that fut gets orig response.
+    assert_eq!(assert_ready_ok!(fut.poll()), "orig-done");
 }
 
-#[test]
-fn hedge_hedge_completes_first() {
-    task::mock(|cx| {
-        clock::mock(|time| {
-            let (mut service, handle) = new_service(TestPolicy);
-            pin_mut!(handle);
+#[tokio::test]
+async fn hedge_hedge_completes_first() {
+    time::pause();
 
-            assert_ready_ok!(service.poll_ready(cx));
-            let fut = service.call("orig");
-            pin_mut!(fut);
-            // Check that orig request has been issued.
-            let _req = assert_request_eq!(handle, "orig");
-            // Check fut is not ready.
-            assert_pending!(fut.as_mut().poll(cx));
+    let (mut service, mut handle) = new_service(TestPolicy);
 
-            // Check hedge has not been issued.
-            assert_pending!(handle.as_mut().poll_request(cx));
-            time.advance(Duration::from_millis(10));
-            // Check fut is not ready.
-            assert_pending!(fut.as_mut().poll(cx));
+    assert_ready_ok!(service.poll_ready());
+    let mut fut = task::spawn(service.call("orig"));
 
-            // Check that the hedge has been issued.
-            let hedge_req = assert_request_eq!(handle, "orig");
-            hedge_req.send_response("hedge-done");
-            // Check that fut gets hedge response.
-            assert_eq!(assert_ready_ok!(fut.as_mut().poll(cx)), "hedge-done");
-        });
-    });
+    // Check that orig request has been issued.
+    let _req = assert_request_eq!(handle, "orig");
+
+    // Check fut is not ready.
+    assert_pending!(fut.poll());
+
+    // Check hedge has not been issued.
+    assert_pending!(handle.poll_request());
+    time::advance(Duration::from_millis(11)).await;
+    // Check fut is not ready.
+    assert_pending!(fut.poll());
+
+    // Check that the hedge has been issued.
+    let hedge_req = assert_request_eq!(handle, "orig");
+    hedge_req.send_response("hedge-done");
+    // Check that fut gets hedge response.
+    assert_eq!(assert_ready_ok!(fut.poll()), "hedge-done");
 }
 
-#[test]
-fn completes_before_hedge() {
-    task::mock(|cx| {
-        clock::mock(|_| {
-            let (mut service, handle) = new_service(TestPolicy);
-            pin_mut!(handle);
+#[tokio::test]
+async fn completes_before_hedge() {
+    let (mut service, mut handle) = new_service(TestPolicy);
 
-            assert_ready_ok!(service.poll_ready(cx));
-            let fut = service.call("orig");
-            pin_mut!(fut);
-            // Check that orig request has been issued.
-            let req = assert_request_eq!(handle, "orig");
-            // Check fut is not ready.
-            assert_pending!(fut.as_mut().poll(cx));
+    assert_ready_ok!(service.poll_ready());
+    let mut fut = task::spawn(service.call("orig"));
 
-            req.send_response("orig-done");
-            // Check hedge has not been issued.
-            assert_pending!(handle.as_mut().poll_request(cx));
-            // Check that fut gets orig response.
-            assert_eq!(assert_ready_ok!(fut.as_mut().poll(cx)), "orig-done");
-        });
-    });
+    // Check that orig request has been issued.
+    let req = assert_request_eq!(handle, "orig");
+    // Check fut is not ready.
+    assert_pending!(fut.poll());
+
+    req.send_response("orig-done");
+    // Check hedge has not been issued.
+    assert_pending!(handle.poll_request());
+    // Check that fut gets orig response.
+    assert_eq!(assert_ready_ok!(fut.poll()), "orig-done");
 }
 
-#[test]
-fn request_not_retyable() {
-    task::mock(|cx| {
-        clock::mock(|time| {
-            let (mut service, handle) = new_service(TestPolicy);
-            pin_mut!(handle);
+#[tokio::test]
+async fn request_not_retyable() {
+    time::pause();
 
-            assert_ready_ok!(service.poll_ready(cx));
-            let fut = service.call(NOT_RETRYABLE);
-            pin_mut!(fut);
-            // Check that orig request has been issued.
-            let req = assert_request_eq!(handle, NOT_RETRYABLE);
-            // Check fut is not ready.
-            assert_pending!(fut.as_mut().poll(cx));
+    let (mut service, mut handle) = new_service(TestPolicy);
 
-            // Check hedge has not been issued.
-            assert_pending!(handle.as_mut().poll_request(cx));
-            time.advance(Duration::from_millis(10));
-            // Check fut is not ready.
-            assert_pending!(fut.as_mut().poll(cx));
-            // Check hedge has not been issued.
-            assert_pending!(handle.as_mut().poll_request(cx));
+    assert_ready_ok!(service.poll_ready());
+    let mut fut = task::spawn(service.call(NOT_RETRYABLE));
 
-            req.send_response("orig-done");
-            // Check that fut gets orig response.
-            assert_eq!(assert_ready_ok!(fut.as_mut().poll(cx)), "orig-done");
-        });
-    });
+    // Check that orig request has been issued.
+    let req = assert_request_eq!(handle, NOT_RETRYABLE);
+    // Check fut is not ready.
+    assert_pending!(fut.poll());
+
+    // Check hedge has not been issued.
+    assert_pending!(handle.poll_request());
+    time::advance(Duration::from_millis(10)).await;
+    // Check fut is not ready.
+    assert_pending!(fut.poll());
+    // Check hedge has not been issued.
+    assert_pending!(handle.poll_request());
+
+    req.send_response("orig-done");
+    // Check that fut gets orig response.
+    assert_eq!(assert_ready_ok!(fut.poll()), "orig-done");
 }
 
-#[test]
-fn request_not_clonable() {
-    task::mock(|cx| {
-        clock::mock(|time| {
-            let (mut service, handle) = new_service(TestPolicy);
-            pin_mut!(handle);
+#[tokio::test]
+async fn request_not_clonable() {
+    time::pause();
 
-            assert_ready_ok!(service.poll_ready(cx));
-            let fut = service.call(NOT_CLONABLE);
-            pin_mut!(fut);
-            // Check that orig request has been issued.
-            let req = assert_request_eq!(handle, NOT_CLONABLE);
-            // Check fut is not ready.
-            assert_pending!(fut.as_mut().poll(cx));
+    let (mut service, mut handle) = new_service(TestPolicy);
 
-            // Check hedge has not been issued.
-            assert_pending!(handle.as_mut().poll_request(cx));
-            time.advance(Duration::from_millis(10));
-            // Check fut is not ready.
-            assert_pending!(fut.as_mut().poll(cx));
-            // Check hedge has not been issued.
-            assert_pending!(handle.as_mut().poll_request(cx));
+    assert_ready_ok!(service.poll_ready());
+    let mut fut = task::spawn(service.call(NOT_CLONABLE));
 
-            req.send_response("orig-done");
-            // Check that fut gets orig response.
-            assert_eq!(assert_ready_ok!(fut.as_mut().poll(cx)), "orig-done");
-        });
-    });
+    // Check that orig request has been issued.
+    let req = assert_request_eq!(handle, NOT_CLONABLE);
+    // Check fut is not ready.
+    assert_pending!(fut.poll());
+
+    // Check hedge has not been issued.
+    assert_pending!(handle.poll_request());
+    time::advance(Duration::from_millis(10)).await;
+    // Check fut is not ready.
+    assert_pending!(fut.poll());
+    // Check hedge has not been issued.
+    assert_pending!(handle.poll_request());
+
+    req.send_response("orig-done");
+    // Check that fut gets orig response.
+    assert_eq!(assert_ready_ok!(fut.poll()), "orig-done");
 }
 
 type Req = &'static str;
@@ -176,7 +157,7 @@ impl tower_hedge::Policy<Req> for TestPolicy {
     }
 }
 
-fn new_service<P: Policy<Req> + Clone>(policy: P) -> (Hedge<Mock, P>, Handle) {
+fn new_service<P: Policy<Req> + Clone>(policy: P) -> (mock::Spawn<Hedge<Mock, P>>, Handle) {
     let (service, handle) = tower_test::mock::pair();
 
     let mock_latencies: [u64; 10] = [1, 1, 1, 1, 1, 1, 1, 1, 10, 10];
@@ -189,5 +170,6 @@ fn new_service<P: Policy<Req> + Clone>(policy: P) -> (Hedge<Mock, P>, Handle) {
         Duration::from_secs(60),
         &mock_latencies,
     );
-    (service, handle)
+
+    (mock::Spawn::new(service), handle)
 }