+++ tower-spawn-ready/src/lib.rs
@@ -1,4 +1,4 @@
-#![doc(html_root_url = "https://docs.rs/tower-spawn-ready/0.3.0-alpha.2")]
+#![doc(html_root_url = "https://docs.rs/tower-spawn-ready/0.3.0")]
 #![warn(
     missing_debug_implementations,
     missing_docs,
@@ -10,7 +10,6 @@
 //! When an underlying service is not ready, drive it to readiness on a
 //! background task.
 
-pub mod error;
 pub mod future;
 mod layer;
 mod make;
@@ -19,3 +18,6 @@ mod service;
 pub use crate::layer::SpawnReadyLayer;
 pub use crate::make::{MakeFuture, MakeSpawnReady};
 pub use crate::service::SpawnReady;
+
+/// Errors produced by `SpawnReady`.
+pub(crate) type Error = Box<dyn std::error::Error + Send + Sync>;