+++ tower-util/src/ready.rs
@@ -1,7 +1,6 @@
 use std::{fmt, marker::PhantomData};
 
 use futures_util::ready;
-use pin_project::pin_project;
 use std::{
     future::Future,
     pin::Pin,
@@ -12,11 +11,13 @@ use tower_service::Service;
 /// Future yielding a `Service` once the service is ready to process a request
 ///
 /// `Ready` values are produced by `ServiceExt::ready`.
-#[pin_project]
 pub struct Ready<'a, T, Request> {
     inner: &'a mut T,
     _p: PhantomData<fn() -> Request>,
 }
+
+// Safety: This is safe because `Services`'s are always `Unpin`.
+impl<'a, T, Request> Unpin for Ready<'a, T, Request> {}
 
 impl<'a, T, Request> Ready<'a, T, Request>
 where
@@ -37,9 +38,8 @@ where
 {
     type Output = Result<(), T::Error>;
 
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let this = self.project();
-        ready!(this.inner.poll_ready(cx))?;
+    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+        ready!(self.inner.poll_ready(cx))?;
 
         Poll::Ready(Ok(()))
     }