+++ tower-spawn-ready/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.3.0 (December 4, 2019)
+
+- Update to `tower-service 0.3`
+- Update to `tokio 0.2`
+- Remove `Executor` usage
+
 # 0.3.0-alpha.2 (September 30, 2019)
 
 - Move to `futures-*-preview 0.3.0-alpha.19`